Support symbolic matrix creation: sym('a', [2 3])
still work-in-progress: lots of things don't work and I'd like to have
it closer to SymPy's MatrixSymbol in the [2, 3] case (already using that
in the [n, m] case with symbolic size).

Probably needs better MatrixExpr doit() support upstream, because of
the "as_explicit()" workaround I'm currently using in python_header.  That
does something horrid to MatrixSymbols with [2 3] size.