stop generating functions like sin.m on install, do it once
Remove the list from .gitignore.  Edit Makefile for this.
generate_functions.py script no longer needs to make the .gitignore
list.  Start tracking all the autogenerated functions.
If we need changes to these, we can rerun the autogen script, but
idea is to simplify installation and lower the barrier for others
to contribute.