sym: change usage of the actual constructor
sym is a main user interface for the class, but the actual
construction is not really user facing (its created on return from
python).  We need a certain "magic" calling interface for the
python code to call (IPC lives outside this class).  A change
was needed to allow arbitrary number of assumptions such as
`sym('x', 'positive', 'integer')`.