size: For now, force returning 1x1 for symbolic sized
This should be reverted at some point.  But for now, I want to
merge the NaN return python stuff and the display code.  Then
think a little more about having size return "NaN x 3".

Octave won't display these properly in `whos` and the GUI
(I tend to get -BIGINT on x86 Linux and 0 on arm Android.)
It looks like Octave 4.0 will not have this addressed so
planning to think about whether I want this.

It also effects what A(2,2) should do when an Matrix has symbolic
size.

Relabel some tests as xtest as this will cause them to fail.

Reopens #159.