display: fix to keep ascii art out of printf statements
Fixes #234.  Not tests needed as a previous commit added a
doctest which covers this.

Basically, don't use printf on the display string.  This is things
like
````
     /x\
  sin|-|
     \3/
````
can render correctly, otherwise the backslashes cause trouble.
Add some warnings to the code because single-line output is
still suseptable.  Easy to fix on Octave using fputs instead of
fprintf.  But then it won't work on Matlab which has no fputs.
OTOH, I don't think single line output is too likely to have
backslashes (famous last words?)

Perhaps I should do a "myfputs()" fcn which is fputs on Octave
and fprintf on Matlab?