isempty(finiteset()) returns true
Seems obvious but it might be dangerous: maybe somewhere in the
code assumes isempty means one dimension of an array is zero (as
is the case for matrices).

For example: `e = finiteset()`, now `size(e)` gives `[1 1]` but
`isempty(e)` is true.

Also: `length(e)` gives 1.  Maybe we need to overload `e` as well.