Improve support for None
Support None (and/or better error messages).  Add tests to None.
We don't change commands were a mathematics matrix is expected
(i.e., in linear algebra routines like svd).  But in many places
matrix is used more as an array: these are changed to support None.

Tempting to use `isinstance(foo, MatrixBase)` but MatrixExpr are
not MatrixBase...  need to be careful.