Tests for escaped-strings were wrong
The string we have in Octave should be passed into to Python without
changing it.  This means we may have to do some escaping for the string
to get through the IPC mechanism: with the current ones, this means
calling "undo_string_escapes".  Invent a quick-n-dirty implementation
of said for matlab.

Fixes #487.