char: gives human-readable string not sympy srepr
This is more useful, e.g., for labelling a plot and I think its
what people would expect to happen if they cast a object to char.
It does mean that `sym(char(x))` doesn't always reproduce `x` with
the fidelity that it used too because previously `x` gave the
implementation (srepr) of the object.  That change is documented in
char.m.

Fixes #375, also effects (or affects) #91.