Fix hex2dec conversions on Octave >= 4.3
Previous code assumed this was a matlab/octave difference but it was
fixed in upstream Octave.  Now we can do the same thing on new Octave
as Matlab.  Unfortunately the previous "arrayfun" approach is very
slow on Octave.  But I found a new approach with "struct2cell" that
is fast enough on Octave and seems to work on Matlab (very limited
tested).