assume: support command form ("assume x y ...")
This is done by adding a new assume.m that is not a method of the sym
class.  This must be the case because in "assume x y real", x and y
are strings.  But why not remove the method "@sym/assume" with the new
"assume.m"?  Well, for now "assume(x, y, 'real')" gives an error, which
is compatible with SMT behaviour.  It also means the assume *method*
does not need to locate the boundary between variables and conditions,
a process which could introduce bugs.  Maybe in some future commit we
will unify the two .m files, but for now they are separate.

Fixes #421.