assume: "assume x ..." applies to contents of var named x
Instead of just using the string 'x' to build a new symbol, "assume x ..."
should get the variable named "x" from the callers workspace.  That
variable could contain a symbol, or a vector, or whatever.  We trust
`@sym/assume` to do the right thing with the contents.  But we still
need to duplicate the "muck around in callers workspace" code.