symfun: update docs to emphasize "syms" more
We don't really want anything that encourages the habit of passing
strings to "sym".  Regarding symfuns, Issue #290 is very much still
a problem when called like "sym('FF(x)')" so that's another reason
to deemphasize it.  In fact, we doctest these problems and encourage
"syms FF(x)" as the way to create symfuns.