charpoly: fix bad test and simplify code
Maybe the convoluted code (Poly->expr->Poly) was needed for older
sympy?  Works on SymPy v1.0 which is the earliest we support.

Remove a bad test which had the variable of the char poly set to a
symbol appearing in the matrix.  Nothing good can come from that;
newer sympy seems to notice and replaces "x" with "_x"; older
sympy just silently subs it in.  But we don't want to have ill-posed
problems as tests.  Reformat other tests.