remove unicode times symbol from code
Upstream talk is that m-file encoding comes from OS not from this
file.  That's nonsense in this case b/c the docstring in this
file is explicitly identified as utf-8.  But anyway, at least the
docstring will only be broken, not the code...  Fixes #941.

do_highbyte_escape.m is not new code: just copied from inst/private.