Fix for DeprecationWarning on newer python
The following warning is generated when calls are made to distutils

octave:3> syms a
Symbolic pkg v2.9.0+: <stdin>:23: DeprecationWarning: The distutils package is deprecated and slated for removal in Python 3.12. Use setuptools or check PEP 632 for potential alternatives

This changes the call from distutils to packaging.  Shamelessly borrowed
from gnuradio.