Fix for DeprecationWarning on Python 3.10
The following warning is generated when calls are made to distutils

octave:3> syms a
Symbolic pkg v2.9.0+: <stdin>:23: DeprecationWarning: The distutils package is deprecated and slated for removal in Python 3.12. Use setuptools or check PEP 632 for potential alternatives

This changes the call from distutils to packaging.  Shamelessly borrowed
from gnuradio.

* Add 'pip install packaging' to continuous integration workflow

- This is necessary as packaging is now required per PR #1061

* Missed adding pip install packaging to th try_stuff job.

* Attempt to get try_stuff working with packaging

* Attempt to get docker to run pip install packaging

* Change import of LooseVersion to use try: except block

Fixes #1061.