Pythonic fails to initialize symbolic variables without a call to type
Fix for pythonic interface failing due to TypeErrors

calls to 'sym x' and 'rho = sym('(1 + sqrt(5))/2')' cause the
interpreter to fail to bring up the python link.  The former causes the
type error on line 49 and the latter with line 58.

change the calls from py.str to py.type(py.str()) and py.int to
py.type(py.int())

Fix for issue #1115