Fallback on SymPy's recently-added version_tuple
This won't help for older SymPy, specifically before 1.9.  But
hopefully once Python 3.12 arrives and disutils is gone we can depend on
on SymPy 1.10.  Worst case, we can copy the `version_tuple`
implementation into our headers.

For now, at least in the case of up-to-date SymPy, it is not necessary
to have `packaging` installed so we do not need to add another Python
dependency.  So this Fixes Issue #1083.